Refactored files in course description
