+++ frontend/src/pages/CourseSelector/courseDescription/LoadingSkeleton/index.jsx
@@ -1,0 +1,3 @@
+import LoadingSkeleton from "./LoadingSkeleton";
+
+export default LoadingSkeleton;
