+++ frontend/src/pages/CourseSelector/courseDescription/LoadingSkeleton/index.less
@@ -1,0 +1,23 @@
+.loading-skeleton-container {
+  display: flex;
+  gap: 5vw;
+  overflow: hidden;
+
+  .loading-skeleton-description {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    width: 50vw;
+    overflow: hidden;
+    gap: 20px;
+  }
+
+  .loading-skeleton-sidebar {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    width: 20vw;
+    gap: 20px;
+  }
+
+}
