+++ frontend/src/pages/CourseSelector/courseDescription/index.jsx
@@ -1,0 +1,3 @@
+import CourseDescription from "./CourseDescription";
+
+export default CourseDescription;
