+++ frontend/src/pages/CourseSelector/courseDescription/index.less
@@ -1,4 +1,3 @@
-
 .cs-description-root {
   padding: 1.6rem 2rem 1.6rem 2rem;
   display: flex;
@@ -14,14 +13,17 @@
   margin-right: 4rem;
   width: 75%;
 }
+
 .cs-course-attr {
   border-bottom: #d9d9d9 solid 1px; //grey-5
   padding-bottom: 10px;
   padding-top: 10px;
 }
+
 .cs-final-attr {
   padding-top: 10px;
 }
+
 .cs-desc-title-bar {
   display: flex;
   flex-direction: row;
@@ -32,7 +34,6 @@
 .cs-alert {
   margin-bottom:  1rem !important;
 }
-
 
 .course-tag-cont {
   display: flex;
@@ -49,12 +50,6 @@
   text-align: center;
 }
 
-.dark {
-  .ant-typography {
-    color: white
-  }
-}
-
 .infographic {
   width: 70%;
   height: 70%;
@@ -62,35 +57,8 @@
 }
 
 
-.addBtn {
-  display: flex;
-  align-items: center;
-  gap: 10px;
+.dark {
+  .ant-typography {
+    color: white
+  }
 }
-
-.skelDescContainer {
-  display: flex;
-  gap: 5vw;
-  overflow: hidden;
-}
-
-.skelDescInfo {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  width: 50vw;
-  overflow: hidden;
-  gap: 20px;
-}
-
-.skelDescSidebar {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  width: 20vw;
-  gap: 20px;
-}
-
-
-
-
