+++ frontend/src/pages/CourseSelector/main.jsx
@@ -3,7 +3,7 @@ import axios from "axios";
 import { Switch } from "antd";
 import { useSelector } from "react-redux";
 import CourseMenu from "./courseMenu/CourseMenu";
-import CourseDescription from "./courseDescription/CourseDescription";
+import CourseDescription from "./courseDescription";
 import CourseTabs from "./courseTabs/CourseTabs";
 import "./main.less";
 import SearchCourse from "./SearchCourse";
